HOP-12
* Move Oracle database to plugin
* Move MySql database to plugin
* Adapate MariaDB dependencies to MySQL database
* Remove Oracle and MysSql test from Greenplum
