+++ assemblies/plugins/databases/pom.xml
@@ -46,7 +46,9 @@
         <module>msaccess-assemblies</module>
         <module>mssqlnative-assemblies</module>
         <module>mssql-assemblies</module>
+        <module>mysql-assemblies</module>
         <module>netezza-assemblies</module>
+        <module>oracle-assemblies</module>
         <module>oraclerdb-assemblies</module>
         <module>remedy-ar-system-assemblies</module>
         <module>sapdb-assemblies</module>
