+++ core/src/test/java/org/apache/hop/core/database/util/DatabaseLogExceptionFactoryTest.java
@@ -22,30 +22,23 @@
 
 package org.apache.hop.core.database.util;
 
-import com.mysql.jdbc.MysqlDataTruncation;
-import com.mysql.jdbc.PacketTooBigException;
-import org.apache.hop.core.database.DatabaseInterface;
-import org.apache.hop.core.database.DatabaseMeta;
-import org.apache.hop.core.database.MySQLDatabaseMeta;
-import org.apache.hop.core.exception.HopDatabaseException;
+import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.mock;
+
 import org.apache.hop.core.logging.LogTableCoreInterface;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-import static org.junit.Assert.assertEquals;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
 public class DatabaseLogExceptionFactoryTest {
 
   private LogTableCoreInterface logTable;
-  private final String SUPPRESSABLE = "org.apache.hop.core.database.util.DatabaseLogExceptionFactory$SuppressBehaviour";
-  private final String THROWABLE = "org.apache.hop.core.database.util.DatabaseLogExceptionFactory$ThrowableBehaviour";
-  private final String
+  private static final String SUPPRESSABLE = "org.apache.hop.core.database.util.DatabaseLogExceptionFactory$SuppressBehaviour";
+  private static final String THROWABLE = "org.apache.hop.core.database.util.DatabaseLogExceptionFactory$ThrowableBehaviour";
+  private static final String
     SUPPRESSABLE_WITH_SHORT_MESSAGE =
     "org.apache.hop.core.database.util.DatabaseLogExceptionFactory$SuppressableWithShortMessage";
-  private final String PROPERTY_VALUE_TRUE = "Y";
+  private static final String PROPERTY_VALUE_TRUE = "Y";
 
   @Before public void setUp() {
     logTable = mock( LogTableCoreInterface.class );
@@ -86,63 +79,6 @@ public class DatabaseLogExceptionFactoryTest {
     assertEquals( THROWABLE, strategyName );
   }
 
-  /**
-   * PDI-5153
-   * Test that in case of PacketTooBigException exception there will be no stack trace in log
-   */
-  @Test
-  public void testExceptionStrategyWithPacketTooBigException() {
-    DatabaseMeta databaseMeta = mock( DatabaseMeta.class );
-    DatabaseInterface databaseInterface = new MySQLDatabaseMeta();
-    PacketTooBigException e = new PacketTooBigException();
-
-    when( logTable.getDatabaseMeta() ).thenReturn( databaseMeta );
-    when( databaseMeta.getDatabaseInterface() ).thenReturn( databaseInterface );
-
-    LogExceptionBehaviourInterface
-      exceptionStrategy =
-      DatabaseLogExceptionFactory.getExceptionStrategy( logTable, new HopDatabaseException( e ) );
-    String strategyName = exceptionStrategy.getClass().getName();
-    assertEquals( SUPPRESSABLE_WITH_SHORT_MESSAGE, strategyName );
-  }
-
-  /**
-   * PDI-5153
-   * Test that in case of MysqlDataTruncation exception there will be no stack trace in log
-   */
-  @Test public void testExceptionStrategyWithMysqlDataTruncationException() {
-    DatabaseMeta databaseMeta = mock( DatabaseMeta.class );
-    DatabaseInterface databaseInterface = new MySQLDatabaseMeta();
-    MysqlDataTruncation e = new MysqlDataTruncation();
-
-    when( logTable.getDatabaseMeta() ).thenReturn( databaseMeta );
-    when( databaseMeta.getDatabaseInterface() ).thenReturn( databaseInterface );
-
-    LogExceptionBehaviourInterface
-      exceptionStrategy =
-      DatabaseLogExceptionFactory.getExceptionStrategy( logTable, new HopDatabaseException( e ) );
-    String strategyName = exceptionStrategy.getClass().getName();
-    assertEquals( SUPPRESSABLE_WITH_SHORT_MESSAGE, strategyName );
-  }
-
-  /**
-   * Property value has priority
-   */
-  @Test public void testExceptionStrategyWithPacketTooBigExceptionPropSetY() {
-    System.setProperty( DatabaseLogExceptionFactory.HOP_GLOBAL_PROP_NAME, PROPERTY_VALUE_TRUE );
-
-    DatabaseMeta databaseMeta = mock( DatabaseMeta.class );
-    DatabaseInterface databaseInterface = new MySQLDatabaseMeta();
-    PacketTooBigException e = new PacketTooBigException();
-
-    when( logTable.getDatabaseMeta() ).thenReturn( databaseMeta );
-    when( databaseMeta.getDatabaseInterface() ).thenReturn( databaseInterface );
-
-    LogExceptionBehaviourInterface
-      exceptionStrategy =
-      DatabaseLogExceptionFactory.getExceptionStrategy( logTable, new HopDatabaseException( e ) );
-    String strategyName = exceptionStrategy.getClass().getName();
-    assertEquals( THROWABLE, strategyName );
-  }
+ 
 
 }
