+++ plugins/databases/mariadb/pom.xml
@@ -17,11 +17,25 @@
 
     <dependencies>
         <dependency>
+        		<groupId>org.hop</groupId>
+        		<artifactId>hop-plugins-databases-mysql</artifactId>        	
+        		<version>${project.version}</version>
+        		<scope>compile</scope>
+     		</dependency>
+    
+        <dependency>
             <groupId>org.mariadb.jdbc</groupId>
             <artifactId>mariadb-java-client</artifactId>
             <version>2.5.1</version>
             <scope>test</scope>
         </dependency>
+        <dependency>
+        		<groupId>org.hop</groupId>
+        		<artifactId>hop-plugins-databases-mysql</artifactId>        	
+        		<version>${project.version}</version>
+        		<type>test-jar</type>
+        		<scope>test</scope>
+      	</dependency>         
     </dependencies>
 
     <dependencyManagement>
@@ -32,6 +46,12 @@
                 <version>2.5.1</version>
                 <scope>test</scope>
             </dependency>
+        		<dependency>
+        			<groupId>org.hop</groupId>
+        			<artifactId>hop-plugins-databases-mysql</artifactId>
+        			<scope>test</scope>
+        			<version>${project.version}</version>
+      			</dependency>            
         </dependencies>
     </dependencyManagement>
