+++ plugins/databases/mysql/src/test/java/com/mysql/jdbc/MysqlDataTruncation.java
